changed data types, added string concat

+++ lib/counter/counter.hpp
@@ -5,6 +5,7 @@
 #include "door_fsm.hpp"
 #include "lcd.hpp"
 #include "ports.h"
+#include "string.h"
 
 class Counter {
  public:
@@ -13,11 +14,11 @@ class Counter {
   void updateCount(const DoorState &state);
   void setMaxCount(uint8_t maxCount);
   void outputToScreen();
-  uint8_t getOutsideCount();
-  uint8_t getInsideCount();
+  int getOutsideCount();
+  int getInsideCount();
 
  private:
-  uint8_t insideCount;
+  int insideCount;
   uint8_t outsideCount;
   uint8_t maxCount;
   unsigned long lastTime;
